refactor(account): move timing & receipt chain hash to separate mods
